Extract native/java_lang.rs from native/common.rs
